Refactor UploadBlindSigs to enable usage in e2e test
